Compare And Pull Academia for helping Professors
CAPA is a robust programmed build to implement the Object Oriented
Principles of Java including Inheritance, Polymorphism and Abstraction.
The purpose of the program is to compare and pull top scores from the
Miami University System and present the student’s name, student id and
class average to the Chair of the University.